+++ src/pat/inject/_injection.scss
@@ -9,7 +9,7 @@
   position: absolute;
   top: 50%;
   left: 50%;
-  content: '\e838';
+  content: "\e838";
   width: 50px;
   height: 50px;
   margin-top: -25px;
