+++ src/pat/inject/history-inject-source.html
@@ -1,36 +1,32 @@
 <!DOCTYPE html>
 <html>
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>Gilgamesh Epos</title>
+  </head>
 
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-  <title>Gilgamesh Epos</title>
-</head>
-
-<body>
-
-  <article class="pat-rich" id="content">
-    <h3>Tablet 1</h3>
-    <p>
-of him who knew the most of all men know;<br>
-who made the journey; heartbroken; reconciled;<br>
-<br>
-who knew the way things were before the Flood,<br>
-the secret things, the mystery; who went<br>
-<br>
-to the end of the earth, and over; who returned,<br>
-and wrote the story on a tablet of stone.<br>
-<br>
-He built Uruk. He built the keeping place<br>
-of Anu and Ishtar. The outer wall<br>
-<br>
-shines in the sun like brightest copper; the inner<br>
-wall is beyond the imagining of kings.<br>
-    </p>
-    <p>
-      <em>— Gilgamesh</em>
-    </p>
-  </article>
-
-</body>
-
+  <body>
+    <article class="pat-rich" id="content">
+      <h3>Tablet 1</h3>
+      <p>
+        of him who knew the most of all men know;<br />
+        who made the journey; heartbroken; reconciled;<br />
+        <br />
+        who knew the way things were before the Flood,<br />
+        the secret things, the mystery; who went<br />
+        <br />
+        to the end of the earth, and over; who returned,<br />
+        and wrote the story on a tablet of stone.<br />
+        <br />
+        He built Uruk. He built the keeping place<br />
+        of Anu and Ishtar. The outer wall<br />
+        <br />
+        shines in the sun like brightest copper; the inner<br />
+        wall is beyond the imagining of kings.<br />
+      </p>
+      <p>
+        <em>— Gilgamesh</em>
+      </p>
+    </article>
+  </body>
 </html>
