+++ src/pat/inject/index.html
@@ -11,7 +11,8 @@
   <meta name="patterns-push-user-id" content="allan_neece" />
   <title>Injection pattern</title>
   <link rel="stylesheet" href="/style/common.css" type="text/css">
-  <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
+    <script src="/bundle-vendors.js" type="text/javascript" charset="utf-8"></script>
+    <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
 
 </head>
 
@@ -312,7 +313,7 @@
       }
 
       70% {
-        -webkit-transform: scale(.9);
+        -webkit-transform: scale(0.9);
       }
 
       100% {
@@ -323,7 +324,7 @@
     @-moz-keyframes bounceIn {
       0% {
         opacity: 0;
-        -moz-transform: scale(.3);
+        -moz-transform: scale(0.3);
       }
 
       50% {
@@ -332,7 +333,7 @@
       }
 
       70% {
-        -moz-transform: scale(.9);
+        -moz-transform: scale(0.9);
       }
 
       100% {
@@ -343,7 +344,7 @@
     @keyframes bounceIn {
       0% {
         opacity: 0;
-        transform: scale(.3);
+        transform: scale(0.3);
       }
 
       50% {
@@ -352,7 +353,7 @@
       }
 
       70% {
-        transform: scale(.9);
+        transform: scale(0.9);
       }
 
       100% {
@@ -360,7 +361,5 @@
       }
     }
   </style>
-
 </body>
-
 </html>
