+++ src/pat/inject/inject-form.html
@@ -1,10 +1,8 @@
 <html>
-<head></head>
-<body>
-
-	<p id="form-submit-target">
-		Here is the text that shows if it worked…
-	</p>
-
-</body>
+  <head></head>
+  <body>
+    <p id="form-submit-target">
+      Here is the text that shows if it worked…
+    </p>
+  </body>
 </html>
